+++ client/scripts/lib/device-detection.js
@@ -1,0 +1,7 @@
+import device from 'current-device';
+
+export { isMobileBrowser };
+
+function isMobileBrowser() {
+  return device.mobile();
+}