+++ client/scripts/lib/sound-player.js
@@ -1,6 +1,7 @@
 'use strict';
 
 import { Howl } from 'howler';
+import { isMobileBrowser } from './device-detection';
 
 export { createSoundPlayer };
 
@@ -9,6 +10,7 @@ function createSoundPlayer(audioData) {
 }
 
 const NO_SOUND_AVAILABLE = 'NOT AVAILABLE';
+const NO_SOUND_AVAILABLE_ON_MOBILE = 'DISABLED ON MOBILE';
 const SOUND_LOADING = 'LOADING...';
 const SOUND_LOADED = 'LOADED';
 const SOUND_LOADING_ERROR = 'ERROR';
@@ -23,6 +25,12 @@ class SoundCategory {
 
     if (audioData.url.length === 0) {
       console.log('NO_SOUND_DATA');
+      return;
+    }
+
+    const isRunningOnMobile = isMobileBrowser();
+    if (isRunningOnMobile) {
+      this.soundState = NO_SOUND_AVAILABLE_ON_MOBILE;
       return;
     }
 
@@ -56,7 +64,7 @@ class SoundCategory {
     if (!this.soundEnabled || !sampleData.sample) {
       return;
     }
-    console.log('play', sampleData)
+    console.log('play', sampleData);
 
     //TODO handle DUCK, GAIN
 
@@ -75,7 +83,7 @@ class SoundCategory {
 
     if (!hasDedicatedChannel || sampleData.gain) {
       const specificVolume = sampleData.gain || 0.5;
-      this.audioSpritePlayer.volume(specificVolume, playId)
+      this.audioSpritePlayer.volume(specificVolume, playId);
     }
 
     if (sampleData.loop) {